# public/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jubilujah Internal API",
    "version": "1.0.0",
    "summary": "Editorial, pipeline, ratings, awards, radio, and admin surface for Jubilujah.",
    "description": "OpenAPI 3.1 contract implementing every endpoint enumerated in Jubilujah Build Spec v1.0 §19 (API Surface — Initial). Authentication is delegated to JubileeInspire SSO (Build Spec §5); the resulting session cookie (`session_id`, HttpOnly/Secure/SameSite=Strict) authorizes all calls. All mutating requests require a CSRF token header (`X-CSRF-Token`). Role-based gates per §5: minimum `content_editor` for read; `production_manager` for pipeline transitions; `radio_producer` for playlist mutations; `admin` for `/admin/*`. Domain schemas live at `/schemas/v1/*.schema.json` and are referenced via `$ref` from this document.",
    "contact": {
      "name": "Gabriel Ungureanu — Stewardship",
      "url": "https://jubilujah.com"
    },
    "license": {
      "name": "Proprietary — Jubilee Music Internal Use Only",
      "identifier": "LicenseRef-Jubilee-Internal"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3119/api",
      "description": "Local development (server.js)"
    },
    {
      "url": "https://jubilujah.com/api",
      "description": "Production stub (gated by JubileeInspire SSO)"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/admin/publish/{type}/{id}": {
      "parameters": [
        { "$ref": "#/components/parameters/RateableTypePath" },
        { "$ref": "#/components/parameters/IdPath" }
      ],
      "post": {
        "tags": ["Admin"],
        "operationId": "triggerPublish",
        "summary": "Trigger CDN publish for an album or song (admin only).",
        "parameters": [{ "$ref": "#/components/parameters/CsrfHeader" }],
        "responses": {
          "202": {
            "description": "Publish queued.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PublishJob" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" }
        }
      }
    },
    "/admin/users": {
      "get": {
        "tags": ["Admin"],
        "operationId": "listUsers",
        "summary": "List local user-role records (admin only).",
        "responses": {
          "200": {
            "description": "User list.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserList" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/admin/users/{id}/roles": {
      "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
      "patch": {
        "tags": ["Admin"],
        "operationId": "updateUserRoles",
        "summary": "Update the cached roles for a user (admin only).",
        "parameters": [{ "$ref": "#/components/parameters/CsrfHeader" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UserRolesPatch" } }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" }
        }
      }
    },
    "/admin/audit": {
      "get": {
        "tags": ["Admin"],
        "operationId": "listAuditEvents",
        "summary": "List audit events since a given ISO-8601 timestamp (admin only).",
        "parameters": [
          { "$ref": "#/components/parameters/SinceQuery" },
          { "$ref": "#/components/parameters/LimitQuery" }
        ],
        "responses": {
          "200": {
            "description": "Audit events.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AuditEventList" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id",
        "description": "HttpOnly + Secure + SameSite=Strict cookie issued by the backend after a successful JubileeInspire SSO exchange (Build Spec §5)."
      }
    },
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "UUID of the resource.",
        "schema": { "type": "string", "format": "uuid" }
      },
...
      "RateableTypePath": {
        "name": "type",
        "in": "path",
        "required": true,
        "description": "Polymorphic object type.",
        "schema": {
          "type": "string",
          "enum": ["song", "album", "artist", "lyrics", "playlist"]
        }
      },
      "CsrfHeader": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": true,
        "description": "Double-submit CSRF token issued at login (Build Spec §20).",
        "schema": { "type": "string", "minLength": 16 }
      },
...
      "SinceQuery": {
        "name": "since",
        "in": "query",
        "required": false,
        "description": "ISO-8601 timestamp; only events at or after this instant are returned.",
        "schema": { "type": "string", "format": "date-time" }
      },
      "LimitQuery": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Page size (default 50, max 200).",
        "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "No valid session cookie present.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "examples": {
              "missingSession": {
                "value": { "error": "unauthorized", "message": "Session required. Re-authenticate via JubileeInspire SSO." }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but role does not satisfy the route's minimum.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "examples": {
              "needsAdmin": {
                "value": { "error": "forbidden", "message": "Role 'admin' required for this operation." }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource does not exist or is not visible to the caller.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "examples": {
              "missing": { "value": { "error": "not_found", "message": "Resource not found." } }
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "Request body failed Pydantic / domain validation.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ValidationError" },
            "examples": {
              "hebrewArticle": {
                "value": {
                  "error": "validation_failed",
                  "message": "Hebrew article rule violation in lyrics.",
                  "details": [
                    { "field": "lyrics.text", "code": "hebrew_article_rule", "hint": "Use 'the Ruach Kodesh' or 'Ruach HaKodesh' — never 'the Ruach HaKodesh'." }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "Error": { "$ref": "/schemas/v1/error.schema.json" },
      "ValidationError": { "$ref": "/schemas/v1/validation-error.schema.json" },
...
      "PublishJob":     { "$ref": "/schemas/v1/admin.publish-job.schema.json" },
      "User":           { "$ref": "/schemas/v1/admin.user.schema.json" },
      "UserList":       { "$ref": "/schemas/v1/admin.user-list.schema.json" },
      "UserRolesPatch": { "$ref": "/schemas/v1/admin.user-roles.patch.schema.json" },
      "AuditEventList": { "$ref": "/schemas/v1/admin.audit-event-list.schema.json" }
    },
...
  }
}